MeshLine.p with dtype=numpy.float_ not int #146

+++ skfem/mesh/mesh_line.py
@@ -24,7 +24,7 @@ class MeshLine(Mesh):
 
     def __init__(self, p=None, t=None, validate=True):
         if p is None and t is None:
-            p = np.array([[0, 1]])
+            p = np.array([[0., 1.]], dtype=np.float_)
         if len(p.shape) == 1:
             p = np.array([p]) 
         self.p = p
